requirements.txt
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
Pillow
numpy==1.24.4
ultralytics
yolov5
fastapi
uvicorn[standard]
onnxruntime
